Move JavaScript object literal extractor to Facebook post extractor strategy v2 as it's the only place where it's used
